+++ whatlies/__init__.py
@@ -1,7 +1,7 @@
 """
 Whatlies library
 """
-__version__ = "0.5.0"
+__version__ = "0.5.1"
 
 from whatlies.embeddingset import EmbeddingSet
 from whatlies.embedding import Embedding